% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_degrees}
\alias{compute_degrees}
\title{Compute node degrees from adjacency matrix}
\usage{
compute_degrees(A)
}
\arguments{
\item{A}{Sparse adjacency matrix.}
}
\value{
Integer vector of node degrees.
}
\description{
Compute node degrees from adjacency matrix
}
\keyword{internal}
