% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{sis_plot_archetypes}
\alias{sis_plot_archetypes}
\title{Plot immune archetypes (post-hoc clustering)}
\usage{
sis_plot_archetypes(fit, coords, K = 4)
}
\arguments{
\item{fit}{A fitted \code{sisHMRF} object.}

\item{coords}{N x 2 matrix of spatial coordinates.}

\item{K}{Number of archetypes (default: 4).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot immune archetypes (post-hoc clustering)
}
