% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{sis_plot_elbo}
\alias{sis_plot_elbo}
\title{Plot ELBO convergence}
\usage{
sis_plot_elbo(fit)
}
\arguments{
\item{fit}{A fitted \code{sisHMRF} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot ELBO convergence
}
