% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{sis_plot_gradient}
\alias{sis_plot_gradient}
\title{Plot immune gradient map}
\usage{
sis_plot_gradient(fit, coords)
}
\arguments{
\item{fit}{A fitted \code{sisHMRF} object.}

\item{coords}{N x 2 matrix of spatial coordinates.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot immune gradient map
}
