% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{sis_plot_loadings}
\alias{sis_plot_loadings}
\title{Plot feature loading heatmap}
\usage{
sis_plot_loadings(fit, feature_names = NULL)
}
\arguments{
\item{fit}{A fitted \code{sisHMRF} object.}

\item{feature_names}{Optional character vector of feature names.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot feature loading heatmap
}
