% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{sis_plot_scores}
\alias{sis_plot_scores}
\title{Plot composite immune scores (IAS / ISS / IBI)}
\usage{
sis_plot_scores(fit, coords)
}
\arguments{
\item{fit}{A fitted \code{sisHMRF} object.}

\item{coords}{N x 2 matrix of spatial coordinates.}
}
\value{
A \code{ggplot} object (3-panel via patchwork) or list.
}
\description{
Plot composite immune scores (IAS / ISS / IBI)
}
