% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{sis_plot_uncertainty}
\alias{sis_plot_uncertainty}
\title{Plot posterior uncertainty (standard error)}
\usage{
sis_plot_uncertainty(fit, coords, dim_idx = 1)
}
\arguments{
\item{fit}{A fitted \code{sisHMRF} object.}

\item{coords}{N x 2 matrix of spatial coordinates.}

\item{dim_idx}{Which latent dimension (default: 1).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot posterior uncertainty (standard error)
}
